[sound] ALSA: ASoC: meson: axg-toddr: fix channel order on g12 platforms
Message-id: <1602843533499326-1070-patchset-jkysela@redhat.com>
Patchwork-id: 332636
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1069/1534] ALSA: ASoC: meson: axg-toddr: fix channel order on g12 platforms
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

On g12 and following platforms, The first channel of record with more than
2 channels ends being placed randomly on an even channel of the output.

On these SoCs, a bit was added to force the first channel to be placed at
the beginning of the output. Apparently the behavior if the bit is not set
is not easily predictable. According to the documentation, this bit is not
present on the axg series.

Set the bit on g12 and fix the problem.

Fixes: a3c23a8ad4dc ("ASoC: meson: axg-toddr: add g12a support")
Reported-by: Nicolas Belin <nbelin@baylibre.com>
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Link: https://lore.kernel.org/r/20200828151438.350974-1-jbrunet@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 9c4b205a20f483d8a5d1208cfec33e339347d4bd)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>