[sound] ALSA: hda/realtek - The Mic on a RedmiBook doesn't work
Message-id: <1602843533499326-1085-patchset-jkysela@redhat.com>
Patchwork-id: 332708
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1084/1534] ALSA: hda/realtek - The Mic on a RedmiBook doesn't work
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The Mic connects to the Nid 0x19, but the configuration of Nid 0x19
is not defined to Mic, and also need to set the coeff to enable the
auto detection on the Nid 0x19. After this change, the Mic plugging
in or plugging out could be detected and could record the sound from
the Mic.

And the coeff value is suggested by Kailang of Realtek.

Cc: Kailang Yang <kailang@realtek.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Hui Wang <hui.wang@canonical.com>
Link: https://lore.kernel.org/r/20200909020041.8967-1-hui.wang@canonical.com
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit fc19d559b0d31b5b831fd468b10d7dadafc0d0ec)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>