[sound] ALSA: ASoC: fsl_sai: Drop TMR/RMR settings for synchronous mode
Message-id: <1602843533499326-1105-patchset-jkysela@redhat.com>
Patchwork-id: 332777
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1104/1534] ALSA: ASoC: fsl_sai: Drop TMR/RMR settings for synchronous mode
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Tx synchronous with Rx: The RMR is the word mask register, it is used
to mask any word in the frame, it is not relating to clock generation,
So it is no need to be changed when Tx is going to be enabled.

Rx synchronous with Tx: The TMR is the word mask register, it is used
to mask any word in the frame, it is not relating to clock generation,
So it is no need to be changed when Rx is going to be enabled.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Nicolin Chen <nicoleotsuka@gmail.com>
Link: https://lore.kernel.org/r/20200805063413.4610-3-shengjiu.wang@nxp.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7b3bee091ec375777ade2a37e4b0c9319f92de27)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>