[sound] ALSA: ASoC: sun8i-codec: Add a quirk for LRCK inversion
Message-id: <1602843533499326-1109-patchset-jkysela@redhat.com>
Patchwork-id: 332742
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1108/1534] ALSA: ASoC: sun8i-codec: Add a quirk for LRCK inversion
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

On the A64, as tested using the PinePhone, the current code causes the
left/right channels to be swapped during I2S playback from the CPU on
AIF1, and breaks DSP_A communication with the modem on AIF2. Both of
these are fixed when LRCK is no longer inverted.

Trusting that the comment in the code is correct, the existing behavior
is kept for the A33.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Link: https://lore.kernel.org/r/20200726012557.38282-5-samuel@sholland.org
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7518805fb636308909a6a7953e9fdb194abb15f8)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>