[sound] ALSA: ASoC: sun50i-codec-analog: Enable DAPM for earpiece switch
Message-id: <1602843533499326-1121-patchset-jkysela@redhat.com>
Patchwork-id: 332772
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1120/1534] ALSA: ASoC: sun50i-codec-analog: Enable DAPM for earpiece switch
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

By including the earpiece mute switch in the DAPM graph, both the
earpiece amplifier and the Mixer/DAC inputs can be powered off when
the earpiece is muted.

While the widget is really just a simple switch, it is represented
as a "mixer with named controls" to avoid including the widget name
in the kcontrol name. Otherwise, it is not possible to give the widget
an accurate, descriptive name without changing the kcontrol name
seen by userspace (which should be stable).

The mute switch is between the source selection and the amplifier,
as per the diagram in the SoC manual.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Link: https://lore.kernel.org/r/20200726025334.59931-9-samuel@sholland.org
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7829e68d55692c9f7f5665ebec9fa1f33d5ad72f)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>