[sound] ALSA: soundwire: intel: reinitialize IP+DSP in .prepare(), but only when resuming
Message-id: <1602843533499326-1143-patchset-jkysela@redhat.com>
Patchwork-id: 332711
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1142/1534] ALSA: soundwire: intel: reinitialize IP+DSP in .prepare(), but only when resuming
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The .prepare() callback is invoked for normal streaming, underflows or
during the system resume transition. In the latter case, the context
for the ALH PDIs is lost, and the DSP is not initialized properly
either, but the bus parameters don't need to be recomputed.

Conversely, when doing a regular .prepare() during an underflow, the
ALH/SHIM registers shall not be changed as the hardware cannot be
reprogrammed after the DMA started (hardware spec requirement).

This patch adds storage of PDI and hw_params in the DAI dma context,
and the difference between the types of .prepare() usages is handled
via a simple boolean, updated when suspending, and tested for in the
.prepare() case.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20200817152923.3259-6-yung-chuan.liao@linux.intel.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit a5a0239c27fe1125826c5cad4dec9cd1fd960d4a)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>