[sound] ALSA: ASoC: hdac_hdmi: tidy up a memset()
Message-id: <1602843533499326-1224-patchset-jkysela@redhat.com>
Patchwork-id: 332843
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1223/1534] ALSA: ASoC: hdac_hdmi: tidy up a memset()
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The ARRAY_SIZE() is the number of the elements but we want to use the
number of bytes.  Fortunately, in this case the value is the same so it
doesn't affect runtime.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Link: https://lore.kernel.org/r/20200825104623.GA278587@mwanda
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 672072976bf0db6e0aca5382bcf03bc90f439923)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>