rhdocs: tighten up Bugzilla specification rules
Bugzilla: INTERNAL

Particularly with our impending move to a gitlab-based workflow, we need
to tighten up the ways in which we specify Bugzilla numbers. Much like
with the commit ID case, we need to be stricter in what we allow as a
valid specifier, to help maintainer and reviewer tooling with recognition
of what is and is not a valid bug number. With merge requests running
through automated webhooks to ferret out and validate bugzillas, we need
to eliminate as many false positives as we can, while still making sure we
catch the valid bugs. This is a heck of a lot easier if we are less
permissive about what we'll consider a valid specifier.

Major changes here: all Bugzilla specifier lines must start with
"Bugzilla:", and these must be contained in the patch description body.
The subject line can optionally still carry the bug number right now,
while we're still emailing patchsets directly to RHKL, but in the near
future, people will be creating Merge Requests on GitLab, where we're
operating on commits in a git branch instead of emails[*], so the
overloaded "[PATCH RHEL-9.3 v2 27/3045 BZ 1234567]" paradigm doesn't
really make sense in a git merge request based workflow. This change
also formalizes the ability to specify multiple bugs on a single line
prefixed with Bugzilla.

I would add an example regex line in perl to this like was done for commit
IDs, but I suck at perl regex, so... if someone can craft it for me, maybe
we can slap it in here.

Related: https://gitlab.com/cki-project/kernel-webhooks/-/merge_requests/55

* = we're actually gunning to have an email bridge that does mail the
contents of your MR to RHKL as we start getting this off the ground, and
those emails might have a subject line generated from the commit body that
does include the bug number in square braces, but you'll have to have the
bug number in the commit body for it to be generated from, and in theory,
the email bridge should go away someday... So get used to it. :)

v2: Make Bugzilla specifier format even stricter, allow multiple bugs on
one line, provide some sample python eval code.

RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>

CC: Jan Stancek <jstancek@redhat.com>
CC: Rafael Aquini <aquini@redhat.com>
CC: Ondrej Mosnacek <omosnace@redhat.com>
CC: Tony Camuso <tcamuso@redhat.com>
CC: Jiri Benc <jbenc@redhat.com>
CC: Vladis Dronov <vdronov@redhat.com>
CC: Herton Krzesinski <herton@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>