[security] selinux: map CAP_PERFMON and CAP_BPF to CAP_SYS_ADMIN
Message-id: <20201022154836.69389-1-yauheni.kaliuta@redhat.com>
Patchwork-id: 333893
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 244/243] selinux: map CAP_PERFMON and CAP_BPF to CAP_SYS_ADMIN
Bugzilla: 1874003
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

From: Ondrej Mosnacek <omosnace@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1874003
Upstream Status: RHEL-only

The v5.8 BPF backport added two new capabilities (CAP_PERFMON and
CAP_BPF), for which the SELinux policy is not adapted yet (not even in
Fedora, unfortunately).

While all uses of the new capabilities have a fallback to CAP_SYS_ADMIN,
due to the way capabilities interact with SELinux, we'd either have a
security regression (things which required CAP_SYS_ADMIN before and now
require CAP_{PERMON|BPF} || CAP_SYS_ADMIN would be always allowed at the
SELinux level if the process has CAP_{PERMON|BPF} in its capability
set) or, in case of MLS policy, we'd get AVC denials for perfmon/bpf in
the audit log (in this case the CAP_SYS_ADMIN fallback would be allowed
or denied as before, so otherwise no difference in functionality).

To work around this and allow things to work the same with the old
policy, which still assumes that only CAP_SYS_ADMIN is needed, this
patch makes SELinux map the PERFMON and BPF capabilities to SYS_ADMIN
internally. That preserves backward compatibility while allowing the use
of the new capabilities by userspace.

It is not a perfect solution, but seems to be the only viable option for
now. Hopefully by RHEL-9 we'll have the policy in Fedora adapted to the
new caps and won't need to carry this patch there.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>