[scsi] scsi: lpfc: Fix retry of PRLI when status indicates its unsupported
Message-id: <20201102183934.26793-38-dkennedy@redhat.com>
Patchwork-id: 336846
Patchwork-instance: patchwork
O-Subject: [RHEL-8.4 PATCH-v2 e-stor 37/58] scsi: lpfc: Fix retry of PRLI when status indicates its unsupported
Bugzilla: 1887549
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Dick Kennedy <dick.kennedy@broadcom.com>

With port bounce/address swaps and timing between initiator GID queries vs
remote port FC4 support registrations, the driver may be in a situation
where it sends PRLIs for both FCP and NVME even though the target may not
support one of the protocols. In this case, the remote port will reject the
PRLI and usually indicate it does not support the request. However, the
driver currently ignores the status of the failure and immediately retries
the PRLI, which is pointless. In the case of this one remote port, the
reception of the PRLI retry caused it to decide to send a LOGO.  The LOGO
restarted the process and the same results happened. It made the remote
port undiscoverable to either protocol.

Add logic to detect the non-support status and not attempt the retry
of the PRLI.

Link: https://lore.kernel.org/r/20200803210229.23063-6-jsmart2021@gmail.com
Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 678768da988081a922b0645b9bd552f6cdd982cc)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>