[fs] vfs: subtype handling moved to fuse
Message-id: <20201103160214.1298244-3-vgoyal@redhat.com>
Patchwork-id: 337255
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 V2 PATCH v2 02/31] vfs: subtype handling moved to fuse
Bugzilla: 1783467
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>

BZ:1783467

commit c7eb6869632a5d33b41d0a00d683b8395392b7ee
Author: David Howells <dhowells@redhat.com>
Date: Mon, 25 Mar 2019 16:38:31 +0000

The unused vfs code can be removed.  Don't pass empty subtype (same as if
->parse callback isn't called).

The bits that are left involve determining whether it's permitted to split the
filesystem type string passed in to mount(2).  Consequently, this means that we
cannot get rid of the FS_HAS_SUBTYPE flag unless we define that a type string
with a dot in it always indicates a subtype specification.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>