[uapi] fuse: reserve values for mapping protocol
Message-id: <20201103160214.1298244-11-vgoyal@redhat.com>
Patchwork-id: 337251
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 V2 PATCH v2 10/31] fuse: reserve values for mapping protocol
Bugzilla: 1783467
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>

BZ:1783467

commit c4bb667eaf520f21b3a3db0489682becc9c49bcc
Author: "Dr. David Alan Gilbert" <dgilbert@redhat.com>
Date: Fri, 2 Aug 2019 18:15:19 +0100

SETUPMAPPING is a command for use with 'virtiofsd', a fuse-over-virtio
implementation; it may find use in other fuse impelementations as well in
which the kernel does not have access to the address space of the daemon
directly.

A SETUPMAPPING operation causes a section of a file to be mapped into a
memory window visible to the kernel.  The offsets in the file and the
window are defined by the kernel performing the operation.

The daemon may reject the request, for reasons including permissions and
limited resources.

When a request perfectly overlaps a previous mapping, the previous mapping
is replaced.  When a mapping partially overlaps a previous mapping, the
previous mapping is split into one or two smaller mappings.

REMOVEMAPPING is the complement to SETUPMAPPING; it unmaps a range of
mapped files from the window visible to the kernel.

The map_alignment field communicates the alignment constraint for
FUSE_SETUPMAPPING/FUSE_REMOVEMAPPING and allows the daemon to constrain the
addresses and file offsets chosen by the kernel.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>