[fs] virtiofs: add logic to free up a memory range
Message-id: <20201103160214.1298244-30-vgoyal@redhat.com>
Patchwork-id: 337247
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 V2 PATCH v2 29/31] virtiofs: add logic to free up a memory range
Bugzilla: 1783467
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>

BZ:1783467

commit 9a752d18c85ae5da28e4a07d52adfd95eacb2495
Author: Vivek Goyal <vgoyal@redhat.com>
Date: Wed, 19 Aug 2020 18:19:56 -0400

Add logic to free up a busy memory range. Freed memory range will be
returned to free pool. Add a worker which can be started to select
and free some busy memory ranges.

Process can also steal one of its busy dax ranges if free range is not
available. I will refer it to as direct reclaim.

If free range is not available and nothing can't be stolen from same
inode, caller waits on a waitq for free range to become available.

For reclaiming a range, as of now we need to hold following locks in
specified order.

	down_write(&fi->i_mmap_sem);
	down_write(&fi->dax->sem);

We look for a free range in following order.

A. Try to get a free range.
B. If not, try direct reclaim.
C. If not, wait for a memory range to become free

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>