[x86] kvm/hyper-v: Synic default SCONTROL MSR needs to be enabled
Message-id: <20201105173657.478449-1-vkuznets@redhat.com>
Patchwork-id: 337773
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt PATCH v2 147/340] x86/kvm/hyper-v: Synic default SCONTROL MSR needs to be enabled
Bugzilla: 1882793
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jon Doron <arilou@gmail.com>

Based on an analysis of the HyperV firmwares (Gen1 and Gen2) it seems
like the SCONTROL is not being set to the ENABLED state as like we have
thought.

Also from a test done by Vitaly Kuznetsov, running a nested HyperV it
was concluded that the first access to the SCONTROL MSR with a read
resulted with the value of 0x1, aka HV_SYNIC_CONTROL_ENABLE.

It's important to note that this diverges from the value states in the
HyperV TLFS of 0.

Signed-off-by: Jon Doron <arilou@gmail.com>
Message-Id: <20200717125238.1103096-2-arilou@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 99b48ecc8e800983b6e00a2350daaeceba1f7406)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>