[x86] kvm: x86: Rename "shared_msrs" to "user_return_msrs"
Message-id: <20201105175435.485614-1-vkuznets@redhat.com>
Patchwork-id: 337895
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt PATCH v2 262/340] KVM: x86: Rename "shared_msrs" to "user_return_msrs"
Bugzilla: 1882793
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Sean Christopherson <sean.j.christopherson@intel.com>

Rename the "shared_msrs" mechanism, which is used to defer restoring
MSRs that are only consumed when running in userspace, to a more banal
but less likely to be confusing "user_return_msrs".

The "shared" nomenclature is confusing as it's not obvious who is
sharing what, e.g. reasonable interpretations are that the guest value
is shared by vCPUs in a VM, or that the MSR value is shared/common to
guest and host, both of which are wrong.

"shared" is also misleading as the MSR value (in hardware) is not
guaranteed to be shared/reused between VMs (if that's indeed the correct
interpretation of the name), as the ability to share values between VMs
is simply a side effect (albiet a very nice side effect) of deferring
restoration of the host value until returning from userspace.

"user_return" avoids the above confusion by describing the mechanism
itself instead of its effects.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Message-Id: <20200923180409.32255-2-sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 7e34fbd05c6350b30161be84d4f8954c9d321292)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>