[x86] perf/x86/rapl: Flip logic on default events visibility
Message-id: <20201106141119.29144-33-mpetlan@redhat.com>
Patchwork-id: 338007
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 032/329] perf/x86/rapl: Flip logic on default events visibility
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 2a3e3f73a23b4ff2c0065d3a42edc18ad94b7851
Author: Stephane Eranian <eranian@google.com>
Date: Wed May 27 15:46:57 2020 -0700

description
===========
This patch modifies the default visibility of the attribute_group
for each RAPL event. By default if the grp.is_visible field is NULL,
sysfs considers that it must display the attribute group.
If the field is not NULL (callback function), then the return value
of the callback determines the visibility (0 = not visible). The RAPL
attribute groups had the field set to NULL, meaning that unless they
failed the probing from perf_msr_probe(), they would be visible. We want
to avoid having to specify attribute groups that are not supported by the HW
in the rapl_msrs[] array, they don't have an MSR address to begin with.

Therefore, we intialize the visible field of all RAPL attribute groups
to a callback that returns 0. If the RAPL msr goes through probing
and succeeds the is_visible field will be set back to NULL (visible).
If the probing fails the field is set to a callback that return 0 (not visible).
Signed-off-by: Michael Petlan <mpetlan@redhat.com>

Signed-off-by: Jan Stancek <jstancek@redhat.com>