[tools] perf hist: Add fast path for duplicate entries check
Message-id: <20201106141119.29144-86-mpetlan@redhat.com>
Patchwork-id: 338062
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 085/329] perf hist: Add fast path for duplicate entries check
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 12e89e65f446476951f42aedeef56b6bd6f7f1e6
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Thu Mar 19 13:25:17 2020 -0700

description
===========
Perf checks the duplicate entries in a callchain before adding an entry.
However the check is very slow especially with deeper call stack.
Almost ~50% elapsed time of perf report is spent on the check when the
call stack is always depth of 32.

The hist_entry__cmp() is used to compare the new entry with the old
entries. It will go through all the available sorts in the sort_list,
and call the specific cmp of each sort, which is very slow.

Actually, for most cases, there are no duplicate entries in callchain.
The symbols are usually different. It's much faster to do a quick check
for symbols first. Only do the full cmp when the symbols are exactly the
same.

The quick check is only to check symbols, not dso. Export
_sort__sym_cmp.

  $ perf record --call-graph lbr ./tchain_edit_64

  Without the patch
  $time perf report --stdio
  real    0m21.142s
  user    0m21.110s
  sys     0m0.033s

  With the patch
  $time perf report --stdio
  real    0m10.977s
  user    0m10.948s
  sys     0m0.027s
Signed-off-by: Michael Petlan <mpetlan@redhat.com>

Signed-off-by: Jan Stancek <jstancek@redhat.com>