[tools] perf stat: Fail on extra comma while parsing events
Message-id: <20201106141119.29144-228-mpetlan@redhat.com>
Patchwork-id: 338330
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 227/329] perf stat: Fail on extra comma while parsing events
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit ae7626418d71ae551fd1f0a85ae7ad9d8b2d4f28
Author: Jiri Olsa <jolsa@kernel.org>
Date: Wed May 20 09:40:50 2020 +0200

description
===========
Ian reported that we allow to parse following:

  $ perf stat -e ,cycles true

which is wrong and we should fail, like we do with this fix:

  $ perf stat -e ,cycles true
  event syntax error: ',cycles'
                        \___ parser error

The reason is that we don't have rule for ',' in 'event' start condition
and it's matched and accepted by default rule.

Add scanner debug support (that Ian already added for expr code),
which was really useful for finding this. It's enabled together with
bison debug via 'make PARSER_DEBUG=1'.
Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>