[tools] perf expr: Allow numbers to be followed by a dot
Message-id: <20201106141119.29144-230-mpetlan@redhat.com>
Patchwork-id: 338326
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 229/329] perf expr: Allow numbers to be followed by a dot
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit a45badc7392b4613708748f6a2cc0c6f63a1c8d6
Author: Ian Rogers <irogers@google.com>
Date: Tue May 19 23:36:52 2020 -0700

description
===========
Metrics like UNC_M_POWER_SELF_REFRESH encode 100 as "100." and
consequently the 100 is treated as a symbol. Alter the regular
expression to allow the dot to be before or after the number.

Note, this passed the pmu-events test as that tests the validity of a
number using strtod rather than lex code. strtod allows the dot after.

Add a test for this behavior.

Fixes: 26226a97724d (perf expr: Move expr lexer to flex)
Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>