[tools] perf metricgroup: Order event groups by size
Message-id: <20201106141119.29144-247-mpetlan@redhat.com>
Patchwork-id: 338344
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 246/329] perf metricgroup: Order event groups by size
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 6bf2102bec4e600e72f87569d2c8e01a3efb340e
Author: Ian Rogers <irogers@google.com>
Date: Wed May 20 11:20:08 2020 -0700

description
===========
When adding event groups to the group list, insert them in size order.
This performs an insertion sort on the group list. By placing the
largest groups at the front of the group list it is possible to see if a
larger group contains the same events as a later group. This can make
the later group redundant - it can reuse the events from the large
group.  A later patch will add this sharing.
Signed-off-by: Michael Petlan <mpetlan@redhat.com>

Signed-off-by: Jan Stancek <jstancek@redhat.com>