[tools] perf metricgroup: Add options to not group or merge
Message-id: <20201106141119.29144-249-mpetlan@redhat.com>
Patchwork-id: 338346
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 248/329] perf metricgroup: Add options to not group or merge
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 05530a7921c0f5149a01e34c4e031c5b18bdc1cc
Author: Ian Rogers <irogers@google.com>
Date: Wed May 20 11:20:10 2020 -0700

description
===========
Add --metric-no-group that causes all events within metrics to not be
grouped. This can allow the event to get more time when multiplexed, but
may also lower accuracy.
Add --metric-no-merge option. By default events in different metrics may
be shared if the group of events for one metric is the same or larger
than that of the second. Sharing may increase or lower accuracy and so
is now configurable.
Signed-off-by: Michael Petlan <mpetlan@redhat.com>

Signed-off-by: Jan Stancek <jstancek@redhat.com>