[tools] perf build: Add NO_LIBCRYPTO=1 to the default set of build tests
Message-id: <20201106141119.29144-273-mpetlan@redhat.com>
Patchwork-id: 338370
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 272/329] perf build: Add NO_LIBCRYPTO=1 to the default set of build tests
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 69fbadbe987a7f9abbc8ac457a1057ebb98e21db
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri May 29 11:38:41 2020 -0300

description
===========
We forgot to add it, so one would have to explicitely ask for it to be
run, fix that by adding it to the set of tests that are performed by
default when one does:

  $ make -C tools/perf build-test

It was being exercised only in the make_minimal test, this patch makes
it be tested in isolation, i.e. disabling only this feature.

Fixes: 8ee4646038e4 ("perf build: Add libcrypto feature detection")
Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>